pyproject.toml
[project]
name = "opportunity-radar"
version = "0.1.0"
description = "AI-powered opportunity scoring engine for creators and indie makers"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "anthropic>=0.96.0",
    "apscheduler>=3.10",
    "asyncpg>=0.31.0",
    "cachetools>=7.0.6",
    "fastapi>=0.136.0",
    "google-api-python-client>=2.194.0",
    "google-api-python-client-stubs>=1.35.0",
    "httpx>=0.28.1",
    "praw>=7.8.1",
    "pydantic-settings>=2.14.0",
    "pytrends>=4.9.2",
    "serpapi>=1.0.2",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn>=0.45.0",
]

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.1",
    "ruff>=0.15.11",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
